test/regression/requirements.txt
pytest
Pillow
pyautogui
opencv-python
